# packages/openapi/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Poveroh API",
        "version": "1.0.0",
        "description": "OpenAPI definition for the Poveroh v1 API"
    },
    "servers": [
        {
            "url": "http://localhost:3001/v1",
            "description": "API Poveroh Localhost v1"
        },
        {
            "url": "http://api.poveroh.local/v1",
            "description": "API Poveroh Localhost proxy v1"
        },
        {
            "url": "http://api.poveroh.com/v1",
            "description": "API Poveroh Production v1"
        }
    ],
    "components": {
        "schemas": {
            "TransactionActionEnum": {
                "type": "string",
                "enum": [
                    "EXPENSES",
                    "INCOME",
                    "TRANSFER"
                ]
            },
            "Subcategory": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "categoryId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "logoIcon": {
                        "type": "string",
                        "format": "uri"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "categoryId",
                    "title",
                    "logoIcon",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
            "GetCategoryListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryDataResponse"
                        },
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "GetCategoryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/CategoryDataResponse"
                            },
                            {
                                "description": "Response data"
                            }
                        ]
                    }
                },
                "required": [
                    "success"
                ]
            },
            "CreateCategoryRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "for": {
                        "$ref": "#/components/schemas/TransactionActionEnum"
                    },
                    "logoIcon": {
                        "type": "string",
                        "format": "uri"
                    },
                    "color": {
                        "type": "string",
                        "default": "#8B5CF6"
                    }
                },
                "required": [
                    "title",
                    "for",
                    "logoIcon"
                ]
            },
            "CreateCategoryMultipartRequest": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/CreateCategoryRequest"
                    },
                    "file": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary",
                            "description": "Optional file upload"
                        }
                    }
                },
                "required": [
                    "file"
                ]
            },
            "CreateCategoryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/CategoryDataResponse"
                            },
                            {
                                "description": "Response data"
                            }
                        ]
                    }
                },
                "required": [
                    "success"
                ]
            },
            "UpdateCategoryRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "for": {
                        "$ref": "#/components/schemas/TransactionActionEnum"
                    },
                    "logoIcon": {
                        "type": "string",
                        "format": "uri"
                    },
                    "color": {
                        "type": "string",
                        "default": "#8B5CF6"
                    }
                }
            },
            "UpdateCategoryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "nullable": true,
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "DeleteCategoryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "nullable": true,
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "CategoryParamsId": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "id"
                ]
            },
            "StringFilter": {
                "type": "object",
                "properties": {
                    "equals": {
                        "type": "string"
                    },
                    "contains": {
                        "type": "string"
                    }
                }
            },
            "CategoryFilters": {
                "type": "object",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/CategoryParamsId"
                    },
                    "title": {
                        "$ref": "#/components/schemas/StringFilter"
                    },
                    "description": {
                        "$ref": "#/components/schemas/StringFilter"
                    },
                    "for": {
                        "$ref": "#/components/schemas/TransactionActionEnum"
                    }
                }
            },
            "FilterOptions": {
                "type": "object",
                "properties": {
                    "skip": {
                        "type": "integer",
                        "minimum": 0,
                        "default": 0
                    },
                    "take": {
                        "type": "integer",
                        "minimum": 0,
                        "exclusiveMinimum": true,
                        "default": 20
                    },
                    "sortBy": {
                        "type": "string",
                        "default": "createdAt"
                    },
                    "sortOrder": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "default": "desc"
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            false
                        ],
                        "description": "Always false for error responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Error message describing what went wrong"
                    },
                    "error": {
                        "nullable": true,
                        "description": "Optional additional error details or context"
                    }
                },
                "required": [
                    "success",
                    "message"
                ]
            },
...
            "CategoryDataResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "for": {
                        "$ref": "#/components/schemas/TransactionActionEnum"
                    },
                    "logoIcon": {
                        "type": "string",
                        "format": "uri"
                    },
                    "color": {
                        "type": "string",
                        "default": "#8B5CF6"
                    },
                    "subcategories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Subcategory"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "for",
                    "logoIcon",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "getCategories",
                "summary": "Get all categories",
                "description": "Retrieve a list of all categories associated with the user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/components/schemas/CategoryFilters"
                        },
                        "required": false,
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "$ref": "#/components/schemas/FilterOptions"
                        },
                        "required": false,
                        "name": "options",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetCategoryListResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "operationId": "createCategory",
                "summary": "Create category",
                "description": "Create a new category with the provided data",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Category data to create",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCategoryMultipartRequest"
                            },
                            "encoding": {
                                "data": {
                                    "contentType": "application/json"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateCategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "operationId": "deleteCategories",
                "summary": "Delete all categories",
                "description": "Delete all categories associated with the user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "getCategoryById",
                "summary": "Get category by ID",
                "description": "Retrieve a specific category by its ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetCategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Category"
                ],
                "operationId": "updateCategory",
                "summary": "Update category",
                "description": "Update an existing category with the provided data",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "description": "Category data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateCategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "operationId": "deleteCategory",
                "summary": "Delete category",
                "description": "Delete an existing category with the provided ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCategoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}